fix: the access endpoint does not work for chekcing principals' permissions
The "access" endpoint should return the permissions for a user for a
given application when the user or service account that is requesting
them has the proper permissions.

However, since the "access" endpoint receives so much traffic, loading
permissions for every user that sent a request to the endpoint was
becoming much of a burden. Therefore, loading the users' permissions was
restricted to a few particular cases.

These changes should allow to satisfy the use case we are after without
sacrificing the restrictions that were set for the endpoint.

RHCLOUD-30043